app/requirements.txt
aiodns==2.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backtrader==1.9.76.123
ccxt==1.50.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
Flask-JWT-Extended==4.2.1
Flask-RESTful==0.3.9
flask_cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask==2.0.3
idna==2.10
Jinja2==3.0.1
matplotlib==3.5.1
mpl_finance==0.10.1
multidict==5.1.0
numpy==1.22.3
prettytable==2.1.0
psycopg2-binary
pycares==4.1.2
pycparser==2.20
pyfiglet==0.8.post1
PySocks==1.7.1
python-binance==1.0.15
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.19.2
pytz==2021.1
requests==2.25.1
rich
SQLAlchemy==1.4.19
TA-Lib==0.4.24
tradingview_ta==3.2.10
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
websocket-client==1.3.1
Werkzeug==2.0.1
yarl==1.6.3